Move HerdMemberDetails-related fragments into their own package; change color scheme and some naming
