+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".HerdMemberDetails"
+            android:name=".herd_member_details.HerdMemberDetails"
             android:theme="@style/AppTheme.NoActionBar"></activity>
         <activity android:name=".DownloadDatabase" />
         <activity
