+++ app/src/main/java/com/example/cattlelog/herd_member_details/OverviewFragment.kt
@@ -1,12 +1,11 @@
 package com.example.cattlelog.herd_member_details
 
-import android.content.Context
-import android.net.Uri
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 
 import com.example.cattlelog.R
 
@@ -15,6 +14,14 @@ import com.example.cattlelog.R
  */
 class OverviewFragment : Fragment() {
 
+    private var herdMemberTagNumber: Int = -1
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        herdMemberTagNumber = (activity as HerdMemberDetails).getTagNumber()
+        Toast.makeText(context, "OverviewFragment received: ${herdMemberTagNumber}", Toast.LENGTH_LONG).show()
+    }
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -22,5 +29,4 @@ class OverviewFragment : Fragment() {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_overview, container, false)
     }
-
 }
