+++ app/src/main/res/layout/activity_herd_member_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HerdMemberDetails">
+    tools:context=".herd_member_details.HerdMemberDetails">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
