+++ app/src/main/res/layout/fragment_herd_member_details.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.PlaceholderFragment">
+    tools:context=".herd_member_details.PlaceholderFragment">
 
     <TextView
         android:id="@+id/section_label"
