+++ app/src/main/res/layout/fragment_treatment.xml
@@ -3,5 +3,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".herd_member_details.TreatmentFragment"
-    android:background="@color/cardview_shadow_start_color"/>
+    tools:context=".herd_member_details.TreatmentFragment"/>
