+++ app/src/main/res/layout/herd_overview_row.xml
@@ -11,7 +11,8 @@
         android:layout_height="fill_parent"
         android:layout_weight="1.0"
         android:gravity="center"
-        android:textAlignment="center" />
+        android:textAlignment="center"
+        android:textSize="@dimen/font_base" />
 
     <TextView
         android:id="@+id/herdBirthDate"
@@ -19,7 +20,8 @@
         android:layout_height="fill_parent"
         android:layout_weight="1.0"
         android:gravity="center"
-        android:textAlignment="center" />
+        android:textAlignment="center"
+        android:textSize="@dimen/font_base" />
 
     <TextView
         android:id="@+id/herdBarnName"
@@ -27,6 +29,7 @@
         android:layout_height="fill_parent"
         android:layout_weight="1.0"
         android:gravity="center"
-        android:textAlignment="center" />
+        android:textAlignment="center"
+        android:textSize="@dimen/font_base" />
 
 </LinearLayout>
